requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
autopep8==2.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==4.1.0
black==23.3.0
celery==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.4.0
decorator==5.1.1
Django==4.2.2
django-celery-beat==2.5.0
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-timezone-field==5.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
executing==1.2.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.2
jedi==0.18.2
kombu==5.3.0
matplotlib-inline==0.1.6
more-itertools==9.1.0
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
pluggy==1.0.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.1
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
random-user-agent==1.0.1
redis==4.5.5
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
